Refactor utility functions and add Foursquare API client for improved functionality and backward compatibility
